Support @ as comment character
Some assemblers prefer the @ character instead of # for comments. Most
notably the GNU ARM assembler, as denoted here:
https://sourceware.org/binutils/docs/as/ARM_002dChars.html#ARM_002dChars

The verbose-arm function of the GNU C compiler also uses @ to provide
verbose information.

With this change, # as well as @ can be used, effectively making it
possible to use this program on ARM machines.